requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pymongo
numpy==1.26.1
pandas==2.1.2
scikit-learn==1.3.2
scipy==1.11.4
python-dotenv==1.0.0